requirements.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
Pillow==9.5.0
drf-spectacular==0.26.2
celery==5.2.7



